[ref] Embed Response struct for ea. XResponse type.
